chore: organize files
